switch to entirely CA-based ssh auth
This should be easier for me to maintain and reason about. It's also a
better example of how to configure and secure ssh clients and servers in
general.

Sand now creates its own user and host certificate authorities, and
uses those CAs to create certs for two-way (user<->host) ssh authentication.

Setting up CAs in an actual production SSH environment is a lot trickier than
what sand is doing (sand only expects to run on hardware that you
already have physical access to and thus complete control over), but this
still may be a useful toy example to learn the basic concepts from.